TEST Object works
Basic TEST object behavior:

If no item is equipped, nothing happens.
If an item is equipped, a message saying "Use {item}" appears.
If the player attempts to use the hammer, the TEST object disappears.
If the player attempts to use any other item, a message saying "I can't use that here." is displayed.